refactor: remove updateParticipantStatus from EventController and EventService; add handleParticipantStatus to ParticipationController and routes
